# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Community Quiz API",
        "description": "Community-Based Quiz Management System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/questions/mcq": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create a Multiple Choice Question (MCQ)",
                "operationId": "92267ef074a6d84123c213a7e5930d0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quiz_id",
                                    "question_text",
                                    "options",
                                    "correct_option"
                                ],
                                "properties": {
                                    "quiz_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "question_text": {
                                        "type": "string",
                                        "example": "What is the capital of France?"
                                    },
                                    "options": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Berlin",
                                            "Paris",
                                            "Rome"
                                        ]
                                    },
                                    "correct_option": {
                                        "description": "Index of the correct option in the options array",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "MCQ created successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/questions/true-false": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create a True/False Question",
                "operationId": "83a1ffcf093785f1a24cbee6c4f80125",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quiz_id",
                                    "question_text",
                                    "correct_answer"
                                ],
                                "properties": {
                                    "quiz_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "question_text": {
                                        "type": "string",
                                        "example": "The sky is blue."
                                    },
                                    "correct_answer": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "True/False Question created successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/questions/short-answer": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create a Short Answer Question",
                "operationId": "f9672a8556a02a5df3e96fc20a372635",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quiz_id",
                                    "question_text",
                                    "correct_answer"
                                ],
                                "properties": {
                                    "quiz_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "question_text": {
                                        "type": "string",
                                        "example": "What is the boiling point of water in Celsius?"
                                    },
                                    "correct_answer": {
                                        "type": "string",
                                        "example": "100"
                                    },
                                    "is_manual_grading": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Short Answer Question created successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/questions/{question}/options": {
            "get": {
                "tags": [
                    "Question Options"
                ],
                "summary": "Get all options for a specific question",
                "operationId": "f8faed67b228b89002674ac9626e73cf",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/options": {
            "post": {
                "tags": [
                    "Question Options"
                ],
                "summary": "Create option",
                "operationId": "c1a5e3d1c9370a02ad013294e295a057",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question_id",
                                    "option_text",
                                    "is_correct"
                                ],
                                "properties": {
                                    "question_id": {
                                        "type": "integer"
                                    },
                                    "option_text": {
                                        "type": "string"
                                    },
                                    "is_correct": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/options/{option}": {
            "put": {
                "tags": [
                    "Question Options"
                ],
                "summary": "Update option",
                "operationId": "f2522e109930a8f7103efbea6db2db27",
                "parameters": [
                    {
                        "name": "option",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "option_text": {
                                        "type": "string"
                                    },
                                    "is_correct": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Question Options"
                ],
                "summary": "Delete option",
                "operationId": "289b63557253629ed45be5eb1c5f34ee",
                "parameters": [
                    {
                        "name": "option",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}